# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Correspondence Workspace API",
    "description": "Built-in mailbox, event-driven case automation, connected system workspaces and human review.",
    "version": "0.9.0"
  },
  "paths": {
...
    "/api/cases/{case_id}/assessment": {
      "get": {
        "tags": [
          "Business rules and validation"
        ],
        "summary": "Current Assessment",
        "operationId": "current_assessment_api_cases__case_id__assessment_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentReport"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        }
      }
    },
    "/api/cases/{case_id}/assessments": {
      "post": {
        "tags": [
          "Business rules and validation"
        ],
        "summary": "Save Assessment",
        "operationId": "save_assessment_api_cases__case_id__assessments_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedAssessment"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        }
      },
      "get": {
        "tags": [
          "Business rules and validation"
        ],
        "summary": "Assessment History",
        "operationId": "assessment_history_api_cases__case_id__assessments_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SavedAssessment"
                  },
                  "title": "Response Assessment History Api Cases  Case Id  Assessments Get"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        }
      }
    },
    "/api/cases/{case_id}/validate-response": {
      "post": {
        "tags": [
          "Business rules and validation"
        ],
        "summary": "Response Preflight",
        "operationId": "response_preflight_api_cases__case_id__validate_response_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftCandidate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationReport"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        }
      }
    },
    "/api/cases/{case_id}/completion-check": {
      "get": {
        "tags": [
          "Business rules and validation"
        ],
        "summary": "Completion Preflight",
        "operationId": "completion_preflight_api_cases__case_id__completion_check_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationReport"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Implemented"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgingResult": {
        "properties": {
          "original_received_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Received At"
          },
          "evaluation_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluation At"
          },
          "received_eastern_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Received Eastern Date"
          },
          "evaluation_eastern_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluation Eastern Date"
          },
          "workdays": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workdays"
          },
          "timezone": {
            "type": "string",
            "const": "America/New_York",
            "title": "Timezone",
            "default": "America/New_York"
          },
          "calendar_version": {
            "type": "string",
            "const": "eastern-mon-fri-v1",
            "title": "Calendar Version",
            "default": "eastern-mon-fri-v1"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "original_received_at",
          "evaluation_at",
          "reason"
        ],
        "title": "AgingResult"
      },
...
      "AssessmentCreate": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "expected_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "expected_revision"
        ],
        "title": "AssessmentCreate"
      },
      "AssessmentReport": {
        "properties": {
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "case_revision": {
            "type": "integer",
            "title": "Case Revision"
          },
          "policy_version": {
            "type": "string",
            "title": "Policy Version"
          },
          "evaluation_at": {
            "type": "string",
            "format": "date-time",
            "title": "Evaluation At"
          },
          "input_hash": {
            "type": "string",
            "title": "Input Hash"
          },
          "evidence_versions": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Evidence Versions"
          },
          "client_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Version"
          },
          "routing": {
            "$ref": "#/components/schemas/RoutingResult"
          },
          "classification": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classification"
          },
          "disposition": {
            "type": "string",
            "title": "Disposition"
          },
          "review_required": {
            "type": "boolean",
            "title": "Review Required"
          },
          "authorized_recipient": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authorized Recipient"
          },
          "permitted_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Permitted Actions"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "type": "array",
            "title": "Findings"
          },
          "concerns": {
            "items": {
              "$ref": "#/components/schemas/ConcernPlan"
            },
            "type": "array",
            "title": "Concerns"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskPlan"
            },
            "type": "array",
            "title": "Tasks"
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/AttachmentCheck"
            },
            "type": "array",
            "title": "Attachments"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "case_id",
          "case_revision",
          "policy_version",
          "evaluation_at",
          "input_hash",
          "evidence_versions",
          "client_version",
          "routing",
          "classification",
          "disposition",
          "review_required",
          "authorized_recipient",
          "permitted_actions",
          "findings",
          "concerns",
          "tasks",
          "attachments"
        ],
        "title": "AssessmentReport"
      },
      "AttachmentCheck": {
        "properties": {
          "evidence_id": {
            "type": "string",
            "title": "Evidence Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "evidence_id",
          "title",
          "valid",
          "code",
          "reason"
        ],
        "title": "AttachmentCheck"
      },
...
      "ConcernPlan": {
        "properties": {
          "concern_id": {
            "type": "string",
            "title": "Concern Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "disposition": {
            "type": "string",
            "enum": [
              "resolved",
              "pending_borrower",
              "pending_department",
              "referred"
            ],
            "title": "Disposition"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Ids"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "next_review_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Review At"
          },
          "resume_requirement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Requirement"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "concern_id",
          "description",
          "disposition",
          "reason",
          "evidence_ids"
        ],
        "title": "ConcernPlan"
      },
...
      "ConcernResponse": {
        "properties": {
          "concern_id": {
            "type": "string",
            "title": "Concern Id"
          },
          "disposition": {
            "type": "string",
            "enum": [
              "resolved",
              "pending_borrower",
              "pending_department",
              "referred"
            ],
            "title": "Disposition"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "concern_id",
          "disposition"
        ],
        "title": "ConcernResponse"
      },
...
      "DraftCandidate": {
        "properties": {
          "case_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Case Revision"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version",
            "default": 1
          },
          "response_type": {
            "type": "string",
            "enum": [
              "interim_acknowledgment",
              "information_request",
              "final_resolution",
              "referral",
              "repetitive"
            ],
            "title": "Response Type"
          },
          "client_code": {
            "type": "string",
            "title": "Client Code"
          },
          "client_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Client Version"
          },
          "sender": {
            "type": "string",
            "title": "Sender"
          },
          "recipient": {
            "type": "string",
            "title": "Recipient"
          },
          "evidence_versions": {
            "additionalProperties": {
              "type": "integer",
              "minimum": 1.0
            },
            "type": "object",
            "title": "Evidence Versions"
          },
          "attachment_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Attachment Ids"
          },
          "disclosure_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Disclosure Ids"
          },
          "concerns": {
            "items": {
              "$ref": "#/components/schemas/ConcernResponse"
            },
            "type": "array",
            "title": "Concerns"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/FactualClaim"
            },
            "type": "array",
            "title": "Claims"
          },
          "body": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "case_revision",
          "response_type",
          "client_code",
          "client_version",
          "sender",
          "recipient",
          "evidence_versions",
          "concerns"
        ],
        "title": "DraftCandidate"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Fields"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "FactualClaim": {
        "properties": {
          "field": {
            "type": "string",
            "enum": [
              "loan_identifier",
              "tax_amount_minor",
              "tax_due_date",
              "tax_scheduled_date",
              "tax_status",
              "tax_bill_received_at",
              "tax_paid_at",
              "current_legal_name",
              "bankruptcy_status",
              "eft_intent",
              "task_completed",
              "document_attached"
            ],
            "title": "Field"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "evidence_id": {
            "type": "string",
            "title": "Evidence Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "field",
          "value",
          "evidence_id"
        ],
        "title": "FactualClaim"
      },
      "Finding": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "rule": {
            "type": "string",
            "title": "Rule"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "blocks": {
            "items": {
              "type": "string",
              "enum": [
                "response",
                "update",
                "closure",
                "ordinary_route"
              ]
            },
            "type": "array",
            "title": "Blocks"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "rule",
          "message"
        ],
        "title": "Finding"
      },
...
      "RoutingResult": {
        "properties": {
          "route": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Route"
          },
          "assessment": {
            "type": "string",
            "title": "Assessment"
          },
          "rule": {
            "type": "string",
            "title": "Rule"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "restrictions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Restrictions"
          },
          "required_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Actions"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "type": "array",
            "title": "Findings"
          },
          "aging": {
            "$ref": "#/components/schemas/AgingResult"
          },
          "inputs": {
            "$ref": "#/components/schemas/RoutingInputs"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "route",
          "assessment",
          "rule",
          "reason",
          "restrictions",
          "required_actions",
          "findings",
          "aging",
          "inputs"
        ],
        "title": "RoutingResult"
      },
...
      "SavedAssessment": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "report": {
            "$ref": "#/components/schemas/AssessmentReport"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "created_at",
          "report"
        ],
        "title": "SavedAssessment"
      },
...
      "TaskPlan": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "task_type": {
            "type": "string",
            "title": "Task Type"
          },
          "action": {
            "type": "string",
            "enum": [
              "reuse",
              "inspect_exception"
            ],
            "title": "Action"
          },
          "result_supported": {
            "type": "boolean",
            "title": "Result Supported"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "task_id",
          "task_type",
          "action",
          "result_supported",
          "reason"
        ],
        "title": "TaskPlan"
      },
...
      "ValidationReport": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "policy_version": {
            "type": "string",
            "title": "Policy Version"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "type": "array",
            "title": "Findings"
          },
          "rendered_body": {
            "type": "string",
            "title": "Rendered Body"
          },
          "response_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "interim_acknowledgment",
                  "information_request",
                  "final_resolution",
                  "referral",
                  "repetitive"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Type"
          },
          "resulting_status": {
            "type": "string",
            "title": "Resulting Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "valid",
          "policy_version",
          "findings",
          "rendered_body",
          "resulting_status"
        ],
        "title": "ValidationReport"
      },
...
    }
  }
}